fix TypingGame tests
